refactor: Restructure package organization to clean architecture, add domain layer (Android Refactor)
