+++ app/src/androidTest/java/com/thoughtworks/moments/MainScreenUiTest.kt
@@ -5,9 +5,9 @@ import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.test.onNodeWithTag
 import androidx.compose.ui.test.onNodeWithText
 import androidx.compose.ui.test.performScrollToIndex
-import com.thoughtworks.moments.api.MomentRepository
-import com.thoughtworks.moments.screen.MainScreen
-import com.thoughtworks.moments.viewmodels.MainViewModel
+import com.thoughtworks.moments.data.remote.repository.MomentRepository
+import com.thoughtworks.moments.ui.screen.MainScreen
+import com.thoughtworks.moments.ui.viewmodels.MainViewModel
 import org.junit.Rule
 import org.junit.Test
