+++ app/src/main/java/com/thoughtworks/moments/data/dto/Tweet.kt
@@ -1,4 +1,4 @@
-package com.thoughtworks.moments.api.entry
+package com.thoughtworks.moments.data.dto
 
 import com.squareup.moshi.JsonClass
