+++ app/src/main/java/com/thoughtworks/moments/ui/screen/theme/Theme.kt
@@ -1,6 +1,5 @@
-package com.thoughtworks.moments.ui.theme
+package com.thoughtworks.moments.ui.screen.theme
 
-import android.app.Activity
 import android.os.Build
 import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.material3.MaterialTheme
