+++ app/src/main/java/com/thoughtworks/moments/ui/viewmodels/MainViewModel.kt
@@ -1,10 +1,10 @@
-package com.thoughtworks.moments.viewmodels
+package com.thoughtworks.moments.ui.viewmodels
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.thoughtworks.moments.api.MomentRepository
-import com.thoughtworks.moments.api.entry.Tweet
-import com.thoughtworks.moments.api.entry.User
+import com.thoughtworks.moments.data.remote.repository.MomentRepository
+import com.thoughtworks.moments.data.dto.Tweet
+import com.thoughtworks.moments.data.dto.User
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.launch
